Initial bringup of RIL
This relies on Dazzozo's HuaweiQualcommRIL.java, which he based on
the CM Qualcomm RIL, but with some changes found by decompiling the
RIL class in the Gingerbread stock ROM.

For now the RIL class has to be manually copied into
    frameworks/base/telephony/java/com/android/internal/telephony/

Calls, SMSes, and USSDs are working, but the Phone.apk is crashing
quite often so this seems to need a bit more work (3G not working,
audio routing not reliable, etc).

The ueventd.u8185.rc is actually ripped straight from the stock ROM,
plus a few CM9-specific things (genlock!).