HuaweiQualcommRIL: Updates to reflect Gingerbread RIL blobs
Dazzozo originally decompiled the stock Gingerbread ROM's framework
to build the RIL class, but eventually updated it to work with the
Ascend G300's ICS RIL blobs.  We're not using ICS blobs, so a few
things had to be changed back to work with our GB blobs.

This fixes the problem that the framework didn't know it was connected
to mobile data; rmnet0 had an IP and could use IP addresses, but
DNS wasn't working and there was no 3G/H icon to show the connection.